@@ -1,21 +1,18 @@
 {
   "name": "@boomerang-io/boomerang-worker",
-  "version": "0.0.0",
+  "version": "1.0.0",
   "private": true,
   "repository": "git@github.com:boomerang-io/boomerang.worker.main.git",
   "author": "Tyson Lawrie <twlawrie@us.ibm.com>",
   "license": "MIT",
   "type": "commonjs",
   "engines": {
-    "node": ">=10",
-    "yarn": "1"
+    "node": ">=18"
   },
-  "workspaces": [
-    "packages/*"
-  ],
   "scripts": {
     "changed": "lerna changed",
     "release": "lerna publish",
+    "version": "lerna version",
     "lint": "eslint packages",
     "prettier": "prettier --write \"**/*.{js,json,md,yml}\"",
     "test": "jest",
@@ -24,17 +21,17 @@
     "prepare": "husky install"
   },
   "devDependencies": {
-    "@commitlint/cli": "^17.0.0",
-    "@commitlint/config-conventional": "^17.0.0",
-    "cross-env": "^7.0.2",
-    "eslint": "^8.18.0",
+    "@commitlint/cli": "^19.2.0",
+    "@commitlint/config-conventional": "^19.2.2",
+    "cross-env": "^7.0.3",
+    "eslint": "^9.8.0",
     "eslint-plugin-node": "^11.1.0",
-    "husky": "^8.0.0",
-    "jest": "^28.1.1",
-    "lerna": "^5.1.4",
+    "husky": "^9.1.4",
+    "jest": "^29.7.0",
+    "lerna": "^8.1.7",
     "lerna-changelog": "^2.2.0",
-    "lint-staged": "^13.0.2",
-    "prettier": "^2.7.1"
+    "lint-staged": "^15.2.8",
+    "prettier": "^3.3.3"
   },
   "lint-staged": {
     "*.js": "eslint packages --fix",
