@@ -1,17 +1,16 @@
 {
   "name": "boomerang-worker-template",
-  "version": "0.1.0",
+  "version": "1.1.0",
   "description": "Boomerang Worker Template",
   "repository": "git@github.com:boomerang-io/boomerang.worker.template.git",
-  "author": "Boomerang Joe <>",
+  "author": "Boomerang Joe",
   "type": "commonjs",
   "engines": {
-    "node": ">=10"
+    "node": ">=18"
   },
   "license": "MIT",
   "dependencies": {
-    "@boomerang-io/worker-cli": "latest",
-    "@boomerang-io/worker-core": "latest"
+    "@boomerang-io/worker-cli": "latest"
   },
   "scripts": {
     "dev": "cross-env NODE_ENV=local boomerang-worker-cli",
@@ -23,15 +22,15 @@
     "test:coverage": "jest --coverage"
   },
   "devDependencies": {
-    "@commitlint/cli": "^8.3.5",
-    "@commitlint/config-conventional": "^8.3.4",
-    "cross-env": "^7.0.2",
-    "eslint": "^6.8.0",
+    "@commitlint/cli": "^19.3.0",
+    "@commitlint/config-conventional": "^19.2.2",
+    "cross-env": "^7.0.3",
+    "eslint": "^9.8.0",
     "eslint-plugin-node": "^11.1.0",
-    "husky": "^4.2.5",
-    "jest": "^25.3.0",
-    "lint-staged": "^10.1.3",
-    "prettier": "^2.0.4"
+    "husky": "^9.1.4",
+    "jest": "^29.7.0",
+    "lint-staged": "^15.2.8",
+    "prettier": "^3.3.3"
   },
   "husky": {
     "hooks": {
